+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-15T13:29:06.223Z",
+    "2026-06-16T04:59:01.026Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {