+++ packages/orq-rc/RELEASES.md
@@ -10009,3 +10009,13 @@ Based on:
 - [typescript v4.11.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.22 - packages/orq-rc
+
+## 2026-06-16 05:01:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.23] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.23 - packages/orq-rc